refactor: Profile 관련 피드백 반영
