+++ src/main/resources/application-local.yml
@@ -1,0 +1,9 @@
+spring:
+  datasource:
+    driver-class-name: com.mysql.cj.jdbc.Driver
+    url: jdbc:mysql://localhost/calorie?serverTimezone=Asia/Seoul
+    username: root
+    password: root
+
+file:
+  dir: "C:\\Users\\dudwl\\OneDrive\\바탕 화면\\calorie-counter\\"
